管理者権限追加

+++ app/Http/Controllers/CategoryController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator; // Validatorのインポート
+use Illuminate\Support\Facades\Auth;
+
 
 class CategoryController extends Controller
 {
@@ -13,6 +15,10 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        // 管理者チェック
+        if (!Auth::check() || !Auth::user()->is_admin === 1) {
+            return redirect('/')->with('error', '権限がありません。');
+        }
         $categories = Category::orderBy('created_at', 'asc')->get();
         return view('categories.index', ['categories' => $categories]);
     }
@@ -22,6 +28,10 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        // 管理者チェック
+        if (!Auth::check() || !Auth::user()->is_admin) {
+            return redirect('/')->with('error', '権限がありません。');
+        }
         return view('categories.create');
     }
 
@@ -30,6 +40,11 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        // 管理者チェック
+        if (!Auth::check() || !Auth::user()->is_admin) {
+            return redirect('/')->with('error', '権限がありません。');
+        }
+
         //バリデーション
         $validator = Validator::make($request->all(), [
             'name' => 'required|min:1|max:255',
@@ -63,6 +78,10 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        // 管理者チェック
+        if (!Auth::check() || !Auth::user()->is_admin) {
+            return redirect('/')->with('error', '権限がありません。');
+        }
         return view('categories.edit', ['category' => $category]);
     }
 
@@ -71,6 +90,10 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        // 管理者チェック
+        if (!Auth::check() || !Auth::user()->is_admin) {
+            return redirect('/')->with('error', '権限がありません。');
+        }
         //バリデーション
         $validator = Validator::make($request->all(), [
             'id' => 'required',
@@ -97,6 +120,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        // 管理者チェック
+        if (!Auth::check() || !Auth::user()->is_admin) {
+            return redirect('/')->with('error', '権限がありません。');
+        }
         $category->delete();       //追加
         return redirect('/categories');  //追加
     }
